requirement.txt
Flask==2.2.3
PyPDF2==3.0.0
python-docx==0.8.11
Pillow==10.0.1
pytest==7.0.1

